third_party/WebKit/LayoutTests/fast/files/url-null.html

<!DOCTYPE HTML>
<html>
 <head>
    <script src="../../resources/js-test.js"></script>
 </head>
 <body>
    <div id="description"></div>
    <div id="console"></div>
<script>
description("Test URL methods with null arguments.");

shouldBe("URL.createObjectURL(null)", "null");
evalAndLog("URL.revokeObjectURL(null)");
var successfullyParsed = true;

</script>
 </body>
</html>
